requirements.txt
click==7.1.2
click-config-file==0.6.0
colorzero==1.1
configobj==5.0.6
flake8==3.8.3
gpiozero==1.5.1
importlib-metadata==1.6.1
mccabe==0.6.1
paho-mqtt==1.5.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
RPi.GPIO==0.7.0
six==1.15.0
smbus==1.1.post2
zipp==3.1.0
